requirement-dev.txt
aiocache==0.11.1
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.2.0
aiosqlite==0.17.0
alembic==1.8.1
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
blinker==1.4
Brotli==1.0.9
browser-cookie3==0.11.4
certifi==2022.6.15
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==36.0.2
distlib==0.3.6
dnspython==2.1.0
dominate==2.7.0
filelock==3.8.0
Flask==2.0.3
Flask-Admin==1.6.0
Flask-Bootstrap4==4.0.2
Flask-Migrate==2.6.0
flask-paginate==2022.1.8
Flask-Script==2.0.5
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
frozenlist==1.3.1
greenlet==1.1.3
h11==0.12.0
h2==4.1.0
hpack==4.0.0
httpcore==0.15.0
httpx==0.23.0
hyperframe==6.0.1
identify==2.5.5
idna==3.3
importlib-metadata==3.7.2
itsdangerous==2.1.2
jeepney==0.8.0
Jinja2==3.1.2
kaitaistruct==0.9
keyring==23.8.2
ldap3==2.9.1
loguru==0.6.0
lz4==4.0.2
Mako==1.2.1
MarkupSafe==2.1.1
mitmproxy==8.0.0
msgpack==1.0.4
multidict==6.0.2
nodeenv==1.7.0
passlib==1.7.4
pbkdf2==1.3
platformdirs==2.5.2
pre-commit==2.20.0
protobuf==3.19.4
psycopg2-binary==2.9.3
publicsuffix2==2.20191221
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.15.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyperclip==1.8.2
python-nmap==0.7.1
python3-nmap==1.5.4
PyYAML==6.0
rfc3986==1.5.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
SecretStorage==3.3.3
simplejson==3.17.6
six==1.16.0
sniffio==1.2.0
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.40
sqlmap==1.6.10
tld==0.12.6
toml==0.10.2
tornado==6.2
typing_extensions==4.3.0
urwid==2.1.2
virtualenv==20.16.5
visitor==0.1.3
wapiti3==3.1.3
Werkzeug==2.2.2
wsproto==1.1.0
WTForms==3.0.1
xmltodict==0.13.0
yarl==1.8.1
yaswfp==0.9.3
zipp==3.8.1
zstandard==0.17.0
